Split models into separate modules
Better for readability and long-term maintainability.
In some cases, enum classes are moved into other classes as "inner" classes
