+++ testrail_data_model/model/__init__.py
@@ -1,0 +1,24 @@
+"""
+Module which provides dataclass models for TestRail API objects
+
+Copyright 2022 SiriusXM-Pandora
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+
+from .project import TestRailProject
+from .suite import TestRailSuite
+from .section import TestRailSection
+from .case import TestRailCase
+from .case_field import TestRailCaseField
+from .case_type import TestRailCaseType
