+++ tox.ini
@@ -41,7 +41,7 @@ commands =
   build: python3 setup.py sdist --dist-dir {toxinidir}/dist
   build: /bin/bash -c 'find {toxinidir}/dist -name "*.tar.gz" | xargs pip3 wheel --no-deps --wheel-dir {toxinidir}/wheels'
   py{py,38,39,310}: coverage run -p -m pytest
-  lint: pylint testrail_data_model
+  lint: pylint testrail_data_model setup.py
   coverage: coverage combine
   coverage: coverage html
   coverage: coverage xml
